feat: Remove toc from right side bar, add toc to the page doc
