+++ app/routes/$libraryId/$version.docs.framework.$framework.$.tsx
@@ -1,6 +1,6 @@
 import { seo } from '~/utils/seo'
 import { createFileRoute } from '@tanstack/react-router'
-import { DocContent } from '~/components/DocContent'
+import { Doc } from '~/components/Doc'
 import { loadDocs } from '~/utils/docs'
 import { getBranch, getLibrary } from '~/libraries'
 import { capitalize } from '~/utils/utils'
@@ -44,12 +44,15 @@ function Docs() {
   const branch = getBranch(library, version)
 
   return (
-    <DocContent
+    <Doc
       title={title}
       content={content}
-      library={library}
+      repo={library.repo}
       branch={branch}
       filePath={filePath}
+      colorFrom={library.colorFrom}
+      colorTo={library.colorTo}
+      shouldRenderToc
     />
   )
 }
