+++ app/routes/$libraryId/$version.docs.tsx
@@ -44,6 +44,7 @@ function DocsRoute() {
       version={version === 'latest' ? library.latestVersion : version!}
       colorFrom={library.colorFrom}
       colorTo={library.colorTo}
+      textColor={library.textColor}
       config={config}
       frameworks={library.frameworks}
       versions={library.availableVersions}
